backend/requirements.txt
# Core Django Framework
Django==4.2.16
djangorestframework==3.14.0

# JWT Authentication
djangorestframework-simplejwt==5.3.0

# CORS handling for frontend
django-cors-headers==4.3.1

# Environment variables management
django-environ==0.11.2

# Database drivers (Windows compatible)
psycopg2-binary==2.9.7

# Static files and deployment
gunicorn==21.2.0
whitenoise==6.6.0

# Image handling (Windows pre-built binary)
Pillow>=10.0.0

# Testing framework
pytest==7.4.3
pytest-django==4.5.2
factory-boy==3.3.0

# Additional essential packages
requests>=2.28.0
setuptools>=65.0.0